videogenerator-repo/python-tts-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch>=2.0.0
torchaudio>=2.0.0
chatterbox-tts
pydantic==2.5.3
python-multipart==0.0.6
